Position open panel from right corner of the viewport always (#85)
* Position open panel from right corner of the viewport always

By using right: 0px and width: x the panel is always relative to the
right corner of the viewport, so when resizing browser viewport (window, dev
tool etc.) re-frisk panel follows the right corner automatically.

Fixes #84

* Position the toggle buttons outside of the container div

Now there is no 100% height div visible when the panel is closed, and
user is able to click to the application on the right corner on the
viewport between the buttons.

(cherry picked from commit 23700e01121817ca0123959199f63e79f1d23051)